requirements.txt
requests==2.7.0
beautifulsoup4==4.4.1
pymysql-0.7.4